% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AddModuleScore_JAS}
\alias{AddModuleScore_JAS}
\title{AddModuleScore_JAS}
\usage{
AddModuleScore_JAS(matrix, signatures, method)
}
\arguments{
\item{method}{oddsratio or likelihood}
}
\value{
matrix
}
\description{
AddModuleScore_JAS
}
\examples{


}
