% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Addsingscore}
\alias{Addsingscore}
\title{Addsingscore}
\usage{
Addsingscore(matrix, signatures)
}
\arguments{
\item{matrix}{single cell expression matrix}

\item{signatures}{gmt file path}
}
\value{
matrix
}
\description{
Addsingscore
}
\examples{


}
