% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{JASMINE}
\alias{JASMINE}
\title{JASMINE: Signature Scoring via JASMINE mergining Means and Enrichment}
\usage{
JASMINE(data, genes, method)
}
\arguments{
\item{method}{oddsratio or likelihood}
}
\value{
JAS_Scores
}
\description{
JASMINE: Signature Scoring via JASMINE mergining Means and Enrichment
}
\examples{


}
