% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{LikelihoodCalculation}
\alias{LikelihoodCalculation}
\title{LikelihoodCalculation: Calculating enrichment of signature genes across each cell (using Likelihood ratio)}
\usage{
LikelihoodCalculation(data, genes)
}
\arguments{
\item{data}{single cell expression matrix}

\item{genes}{}
}
\value{
matrix
}
\description{
LikelihoodCalculation: Calculating enrichment of signature genes across each cell (using Likelihood ratio)
}
\examples{


}
