% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{NormalizationJAS}
\alias{NormalizationJAS}
\title{NormalizationJAS: Scalar [0,1] Normalization of Means and Enrichment across set of cells}
\usage{
NormalizationJAS(JAS_Scores)
}
\arguments{
\item{JAS_Scores}{}
}
\value{
JAS_Scores
}
\description{
NormalizationJAS: Scalar [0,1] Normalization of Means and Enrichment across set of cells
}
\examples{


}
