% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ORCalculation}
\alias{ORCalculation}
\title{ORCalculation: Calculating enrichment of signature genes across each cell 	(using odds ratio)}
\usage{
ORCalculation(data, genes)
}
\arguments{
\item{data}{single cell expression matrix}

\item{genes}{}
}
\value{
matrix
}
\description{
ORCalculation: Calculating enrichment of signature genes across each cell 	(using odds ratio)
}
\examples{


}
