% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RankCalculation}
\alias{RankCalculation}
\title{RankCalculation: Calculating Mean Ranks for signature genes across each cell}
\usage{
RankCalculation(x, genes)
}
\arguments{
\item{x}{single cell expression matrix}

\item{genes}{}
}
\value{
matrix
}
\description{
RankCalculation: Calculating Mean Ranks for signature genes across each cell
}
\examples{


}
