% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_GSVA_pipline}
\alias{run_GSVA_pipline}
\title{run_GSVA_pipline}
\usage{
run_GSVA_pipline(
  rds,
  signatures,
  method = "ssgsea",
  chunk.size = 1000,
  assay = NULL,
  slot = "counts",
  kcdf = "Poisson",
  BPPARAM = NULL,
  ncores = 1
)
}
\arguments{
\item{rds}{seurat object}

\item{signatures}{gmt file path}

\item{method}{gsva method,'gsva', 'ssgsea', 'zscore', 'plage'}

\item{chunk.size}{How many cells to include in each sub-matrix}

\item{assay}{Pull out data from this assay of the Seurat object,(if NULL, use \code{DefaultAssay(obj)})}

\item{slot}{Pull out data from this slot of the Seurat object,default is counts}

\item{kcdf}{"Gaussian", "Poisson", "none"   Character string denoting the kernel to use during the non-parametric estimation of the cumulative distribution function of expression levels across samples when ‘method="gsva"’.  By default, ‘kcdf="Gaussian"’ which is suitable when input expression values are continuous, such as microarray fluorescent units in logarithmic scale, RNA-seq log-CPMs, log-RPKMs or log-TPMs.  When input expression values are integer counts, such as those derived from RNA-seq experiments, then this argument should be set to ‘kcdf="Poisson"’.}

\item{BPPARAM}{A [BiocParallel::bpparam()] object that tells UCell how to parallelize. If provided, it overrides the `ncores` parameter.}

\item{ncores}{ncores Number of processors to parallelize computation. If \code{BPPARAM = NULL}, the function uses,\code{BiocParallel::MulticoreParam(workers=ncores)}}
}
\value{
seurat object
}
\description{
run_GSVA_pipline
}
\examples{


}
