% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_JAS_pipline}
\alias{run_JAS_pipline}
\title{run_JAS_pipline}
\usage{
run_JAS_pipline(
  rds,
  slot = "counts",
  assay = NULL,
  chunk.size = 1000,
  signatures,
  BPPARAM = NULL,
  ncores = 1,
  method = "likelihood"
)
}
\arguments{
\item{rds}{seurat object}

\item{slot}{Pull out data from this slot of the Seurat object,default is counts}

\item{assay}{Pull out data from this assay of the Seurat object,(if NULL, use \code{DefaultAssay(obj)})}

\item{chunk.size}{How many cells to include in each sub-matrix}

\item{signatures}{gmt file path}

\item{BPPARAM}{A [BiocParallel::bpparam()] object that tells UCell how to parallelize. If provided, it overrides the `ncores` parameter.}

\item{ncores}{ncores Number of processors to parallelize computation. If \code{BPPARAM = NULL}, the function uses,\code{BiocParallel::MulticoreParam(workers=ncores)}}

\item{method}{oddsratio or likelihood}
}
\value{
seurat object
}
\description{
run_JAS_pipline
}
\examples{


}
