% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scgmt}
\alias{scgmt}
\title{scgmt}
\usage{
scgmt(
  rds,
  slot = NULL,
  assay = NULL,
  chunk.size = 1000,
  signatures,
  BPPARAM = NULL,
  kcdf = NULL,
  ncores = 1,
  method
)
}
\arguments{
\item{rds}{seurat object}

\item{slot}{Pull out data from this slot of the Seurat object,default is counts}

\item{assay}{Pull out data from this assay of the Seurat object}

\item{chunk.size}{How many cells to include in each sub-matrix}

\item{signatures}{gmt file path}

\item{BPPARAM}{A [BiocParallel::bpparam()] object that tells UCell how to parallelize. If provided, it overrides the `ncores` parameter.}

\item{ncores}{ncores Number of processors to parallelize computation.}

\item{method}{"AUCell","UCell","AddModuleScore","gsva","ssgsea","zscore","plage","VISION","JAS_likelihood","JAS_oddsratio","singscore"}
}
\value{
seurat object
}
\description{
scgmt
}
\examples{


}
