% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scgmt_density_plot}
\alias{scgmt_density_plot}
\title{scgmt_density_plot}
\usage{
scgmt_density_plot(rds, signature, group.by = "ident", cols, alpha = 0.7)
}
\arguments{
\item{rds}{seurat object}

\item{alpha}{}
}
\value{
ggplot2 object
}
\description{
scgmt_density_plot
}
\examples{


}
