% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scgmt_heatmap_plot}
\alias{scgmt_heatmap_plot}
\title{scgmt_heatmap_plot}
\usage{
scgmt_heatmap_plot(
  object,
  signatures,
  group.by = "ident",
  cluster_cols = FALSE,
  cluster_rows = FALSE,
  scale = "row",
  group.order = NULL,
  pathway.order = NULL,
  heat_col = NULL,
  border_color = "white",
  ...
)
}
\arguments{
\item{object}{seurat object}

\item{signatures}{gmt file path}

\item{border_color}{}
}
\value{
pheatmap object
}
\description{
scgmt_heatmap_plot
}
\examples{


}
