% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scgmt_hierarchy}
\alias{scgmt_hierarchy}
\title{scgmt_hierarchy}
\usage{
scgmt_hierarchy(rds, signatures, log.scale = TRUE)
}
\arguments{
\item{rds}{seurat object}

\item{log.scale}{scale the hierarchy coordinates in log2 space. Default: T}

\item{signature}{four signature name}
}
\value{
seurat object
}
\description{
scgmt_hierarchy
}
\examples{


}
