% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scgmt_hierarchy_plot}
\alias{scgmt_hierarchy_plot}
\title{scgmt_hierarchy_plot}
\usage{
scgmt_hierarchy_plot(
  rds,
  group.by = NULL,
  cols = NULL,
  pt.size = 1,
  xlabelaxis = 0.3,
  ylabelaxis = 0.05
)
}
\arguments{
\item{rds}{seurat object}

\item{ylabelaxis}{}
}
\value{
ggplot2 object
}
\description{
scgmt_hierarchy_plot
}
\examples{


}
