% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scgmt_line_plot}
\alias{scgmt_line_plot}
\title{scgmt_line_plot}
\usage{
scgmt_line_plot(
  object,
  signatures,
  group.by = NULL,
  x.lab.title = NULL,
  y.lab.title = NULL,
  face = NULL,
  line.size = 1.2
)
}
\arguments{
\item{object}{seurat object}

\item{signatures}{gmt file path}

\item{line.size}{}
}
\value{
ggplot2 object
}
\description{
scgmt_line_plot
}
\examples{


}
