% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scgmt_ridges_plot}
\alias{scgmt_ridges_plot}
\title{scgmt_ridges_plot}
\usage{
scgmt_ridges_plot(rds, signature, group.by = "ident")
}
\arguments{
\item{rds}{seurat object}

\item{group.by}{}
}
\value{
ggplot2 object
}
\description{
scgmt_ridges_plot
}
\examples{


}
