% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scgmt_scatter_plot}
\alias{scgmt_scatter_plot}
\title{scgmt_scatter_plot}
\usage{
scgmt_scatter_plot(
  rds,
  signature.x,
  signature.y,
  cols = NULL,
  group.by = "ident",
  pt.size = 1,
  guideline = TRUE
)
}
\arguments{
\item{rds}{seurat object}

\item{guideline}{}
}
\value{
ggplot2 object
}
\description{
scgmt_scatter_plot
}
\examples{


}
